Replace app local routers with module CLIs
